@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "vue": "^3.0.0"
+    "vue": "v3.2.31"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
